+++ packages/pouchdb/src/config.default.ts
@@ -4,7 +4,7 @@ import { PouchdbConfig } from './types';
 export const pouchdbConfig = (appInfo: AppInfo): IConfig => {
   return {
     pouchdb: {
-      name: 'pouchdb',
+      name: 'electron-pouchdb',
       path: appInfo.HOME,
     } as PouchdbConfig,
   };