+++ packages/pouchdb/src/index.ts
@@ -1,4 +1,4 @@
 export * from './types';
 export * from './pouchdb_service';
 export * from './config.default';
-export * from './configuration';
+export { PouchdbConfiguration as Configuration } from './configuration';